+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/nerou42/network-journal/compare/v0.7.1...HEAD)
+## [Unreleased](https://github.com/nerou42/network-journal/compare/v0.7.2...HEAD)
+
+
+## [0.7.2](https://github.com/nerou42/network-journal/compare/v0.7.1...v0.7.2) - 2026-07-10
+
+### Fixed
+
+- Backwards compatibility to <0.7.0
 
 
 ## [0.7.1](https://github.com/nerou42/network-journal/compare/v0.7.0...v0.7.1) - 2026-07-09